+++ src/scripts/components/selection-controls/select.vue
@@ -195,8 +195,7 @@ export default {
       UI_SELECT,
       $select: null,
       currentOptions: [],
-      selectedValue: this.modelValue,
-      selectMenuFixed: this.fixed
+      selectedValue: this.modelValue
     };
   },
   computed: {
@@ -227,13 +226,12 @@ export default {
         'mdc-menu',
         'mdc-menu-surface',
         {
-          'mdc-menu-surface--fullwidth': this.fullwidth,
-          'mdc-menu-surface--fixed': this.selectMenuFixed
+          'mdc-menu-surface--fullwidth': this.fullwidth
         }
       ];
     },
     style() {
-      return this.el && this.selectMenuFixed
+      return this.el && this.fixed
         ? { width: this.el.dataset.width || 'auto' }
         : {};
     }
@@ -247,10 +245,6 @@ export default {
     options(val) {
       this.init(val);
     }
-  },
-  beforeMount() {
-    // fix(@material-components-web): overflow inside of the component
-    this.selectMenuFixed = this.fixed;
   },
   mounted() {
     this.$select = new MDCSelect(this.el);
@@ -276,6 +270,11 @@ export default {
         }
       });
     });
+
+    // fix(@material-components-web): overflow inside of the component
+    if (this.fixed) {
+      this.$select.menu.setFixedPosition(true);
+    }
 
     this.init();
   },