+++ src/scripts/components/trees/tree.vue
@@ -18,6 +18,7 @@ const UI_TREE = {
     label: 'label',
     value: 'value',
     children: 'children',
+    hasChildren: 'hasChildren',
     isLeaf: 'isLeaf'
   },
   EVENT: {
@@ -65,6 +66,7 @@ export default {
   emits: [UI_TREE.EVENT.CHANGE],
   data() {
     return {
+      $tree: null,
       nodeList: [],
       treeData: {
         dataFormat: Object.assign(UI_TREE.dataFormat, this.dataFormat),
@@ -84,6 +86,9 @@ export default {
   watch: {
     selectedValue(val) {
       this.$emit(UI_TREE.EVENT.CHANGE, val);
+    },
+    data(val) {
+      this.init(val);
     }
   },
   created() {
@@ -94,7 +99,15 @@ export default {
     }
   },
   mounted() {
-    this.nodeList = new MdcTree(this.data, this.treeData);
+    this.$tree = new MdcTree(this.treeData);
+    this.init();
+  },
+  methods: {
+    init(originData = this.data) {
+      if (!this.nodeList.length) {
+        this.nodeList = this.$tree.getData(originData);
+      }
+    }
   }
 };
 </script>