+++ src/scripts/mixins/table-footer.js
@@ -56,10 +56,7 @@ export default {
   },
   methods: {
     tfootCellClassName(data) {
-      let className = [
-        'mdc-data-table__cell',
-        'mdc-theme--on-surface' // fix(@material-components): dark theme
-      ];
+      let className = ['mdc-data-table__cell'];
 
       className = this.setTextAlignClassName(className, data);
       className = this.setCustomClassName(className, data);